WIP - partition staff assignment from my cases
[skip ci]

Jira ticket: CAMS-418
