--- user-interface/src/case-assignment/AssignAttorneyCasesTable.scss
@@ -1,57 +1,0 @@
-@use '../styles/abstracts/colors';
-@use '../styles/abstracts/tables';
-
-table.case-list {
-  .attorney-list {
-    .table-flex-container {
-      display: flex;
-      position: relative;
-      align-content: center;
-
-      .attorney-list-container {
-        flex: 1;
-        padding-right: 20px;
-      }
-      .table-column-toolbar {
-        padding-left: 20px;
-        align-self: center;
-        text-align: right;
-        border-left: 1px solid #dddddd;
-
-        button {
-          padding: 0.75rem 1rem;
-        }
-      }
-    }
-  }
-  td.case-number {
-    min-width: 115px;
-  }
-  td.filing-date {
-    min-width: 130px;
-  }
-  td.case-title-column {
-    white-space: normal;
-  }
-
-  .case-assignment-modal-toggle {
-    width: 6rem;
-  }
-}
-
-@media screen and (width < 1000px) {
-  table.case-list {
-    margin: 0;
-    width: 100%;
-  }
-}
-
-@media screen and (max-width: 800px) {
-  .usa-table-container--scrollable .usa-table.case-list {
-    width: 100%;
-    padding: 50px 0;
-    td.attorney-list {
-      white-space: wrap;
-    }
-  }
-}
