+++ user-interface/src/my-cases/MyCasesResultsHeader.tsx
@@ -1,0 +1,3 @@
+import { SearchResultsHeader } from '@/search/SearchResultsHeader';
+
+export const MyCasesResultsHeader = SearchResultsHeader;
