+++ user-interface/src/my-cases/MyCasesResultsRow.tsx
@@ -1,0 +1,3 @@
+import { SearchResultsRow } from '@/search/SearchResultsRow';
+
+export const MyCasesResultsRow = SearchResultsRow;
