+++ user-interface/src/search-results/SearchResults.scss
