+++ user-interface/src/staff-assignment/StaffAssignmentScreen.scss
@@ -1,0 +1,13 @@
+.screen-heading {
+  display: block;
+  position: relative;
+
+  h1 {
+    display: inline-block;
+    margin-right: 1rem;
+  }
+  .usa-button {
+    position: absolute;
+    margin-top: 1.5rem;
+  }
+}
