+++ user-interface/src/staff-assignment/modal/AssignAttorneyModal.scss
