+++ user-interface/src/staff-assignment/modal/AssignAttorneyModal.test.tsx
@@ -5,12 +5,12 @@ import AssignAttorneyModal, {
   AssignAttorneyModalRef,
 } from './AssignAttorneyModal';
 import React from 'react';
-import { ToggleModalButton } from '../lib/components/uswds/modal/ToggleModalButton';
 import Api from '@/lib/models/api';
 import { MockData } from '@common/cams/test-utilities/mock-data';
 import { CaseBasics } from '@common/cams/cases';
 import { getCamsUserReference } from '@common/cams/session';
 import { MANHATTAN } from '@common/cams/test-utilities/offices.mock';
+import { ToggleModalButton } from '@/lib/components/uswds/modal/ToggleModalButton';
 
 const offices = [MANHATTAN!];
 const susan = MockData.getAttorneyUser({ name: 'Susan Arbeit', offices });
